requirements.txt
numpy>=1.12.0
tensorflow_gpu>=1.0.0
click==6.7
tqdm==4.11.2
PyYAML==3.12
colorlog>=2.10.0
requests==2.13.0
scipy>=0.19.0
tflearn==0.3
